+++ src/main/java/ru/utmn/dyagunov/tax_subsidies/service/TaxSubsidiesServiceInterface.java
@@ -19,5 +19,5 @@ public interface TaxSubsidiesServiceInterface {
 
     void delete(String id);
 
-    Float getAverageObservationValue();
+    Double getAverageObservationValue();
 }
